Replace `AppLibrary`, `SoftwareMap` & associated types with `installedApps: [InstalledApp]` parameters.
Move app uninstall code from deleted `SoftwareMapAppLibrary.swift` to `Uninstall.swift`.

Refactor `SpotlightSoftwareMap.installedApps()` as global `installedApps`.
