+++ Sources/mas/Commands/Install.swift
@@ -23,13 +23,13 @@ extension MAS {
 
         /// Runs the command.
         func run() async throws {
-            try await run(appLibrary: await SoftwareMapAppLibrary(), searcher: ITunesSearchAppStoreSearcher())
+            try await run(installedApps: await installedApps, searcher: ITunesSearchAppStoreSearcher())
         }
 
-        func run(appLibrary: AppLibrary, searcher: AppStoreSearcher) async throws {
+        func run(installedApps: [InstalledApp], searcher: AppStoreSearcher) async throws {
             // Try to download applications with given identifiers and collect results
             let appIDs = appIDs.filter { appID in
-                if let appName = appLibrary.installedApps(withAppID: appID).first?.name, !force {
+                if let appName = installedApps.first(where: { $0.id == appID })?.name, !force {
                     printWarning(appName, "is already installed")
                     return false
                 }
