+++ Sources/mas/Commands/List.swift
@@ -17,11 +17,10 @@ extension MAS {
 
         /// Runs the command.
         func run() async throws {
-            try run(appLibrary: await SoftwareMapAppLibrary())
+            try run(installedApps: await installedApps)
         }
 
-        func run(appLibrary: AppLibrary) throws {
-            let installedApps = appLibrary.installedApps
+        func run(installedApps: [InstalledApp]) throws {
             if installedApps.isEmpty {
                 printError("No installed apps found")
             } else {
