+++ Sources/mas/Commands/Lucky.swift
@@ -29,10 +29,10 @@ extension MAS {
 
         /// Runs the command.
         func run() async throws {
-            try await run(appLibrary: await SoftwareMapAppLibrary(), searcher: ITunesSearchAppStoreSearcher())
+            try await run(installedApps: await installedApps, searcher: ITunesSearchAppStoreSearcher())
         }
 
-        func run(appLibrary: AppLibrary, searcher: AppStoreSearcher) async throws {
+        func run(installedApps: [InstalledApp], searcher: AppStoreSearcher) async throws {
             var appID: AppID?
 
             do {
@@ -52,18 +52,18 @@ extension MAS {
                 fatalError("app ID returned from Apple is null")
             }
 
-            try await install(appID: appID, appLibrary: appLibrary)
+            try await install(appID: appID, installedApps: installedApps)
         }
 
         /// Installs an app.
         ///
         /// - Parameters:
         ///   - appID: App identifier
-        ///   - appLibrary: Library of installed apps
+        ///   - installedApps: List of installed apps
         /// - Throws: Any error that occurs while attempting to install the app.
-        private func install(appID: AppID, appLibrary: AppLibrary) async throws {
+        private func install(appID: AppID, installedApps: [InstalledApp]) async throws {
             // Try to download applications with given identifiers and collect results
-            if let appName = appLibrary.installedApps(withAppID: appID).first?.name, !force {
+            if let appName = installedApps.first(where: { $0.id == appID })?.name, !force {
                 printWarning(appName, "is already installed")
             } else {
                 do {
