--- Sources/mas/Controllers/SoftwareMap.swift
@@ -1,11 +1,0 @@
-//
-//  SoftwareMap.swift
-//  mas
-//
-//  Created by Ben Chatelain on 3/1/20.
-//  Copyright © 2020 mas-cli. All rights reserved.
-//
-
-protocol SoftwareMap {
-    func installedApps() async -> [InstalledApp]
-}
